Added USE_EXISTING_FREEZR and KEEP_FAILED_FREEZR env settings.

+++ tests/freeze-thaw-aws.sh
@@ -10,9 +10,11 @@ region="${4-${AWS_DEFAULT_REGION-us-east-1}}"
 LOG_PREFIX=${LOG_PREFIX-}
 LOG_SUFFIX=${LOG_SUFFIX-}
 pids=()
+pidfile=freeze-thaw-aws.pid
 
 if [ -z "$access_key" -o -z "$secret_key" ]; then
-    echo "Usage: $0 [AWS-ACCESS-KEY-ID [AWS-SECRET-ACCESS-KEY [KEY-NAME [REGION]]]]
+    echo "Usage: $0 [AWS-ACCESS-KEY-ID [AWS-SECRET-ACCESS-KEY \
+[KEY-NAME [REGION]]]]
 
 You can also specify these values through environmental variables
 AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION." >&2
@@ -53,11 +55,14 @@ function terminate {
     # 	pid=$(pidof "$pid")
     for pid in $(allpids); do
 	echo -n "$pid "
-	(kill -- -$pid; kill $pid; sleep 1; kill -9 -- -$pid; kill -9 $pid) >/dev/null 2>/dev/null
+	(kill -- -$pid; kill $pid; sleep 1;
+	    kill -9 -- -$pid; kill -9 $pid) >/dev/null 2>/dev/null
     done
 
     echo "done"
     pids=()
+    rm -f $pidfile
+
     exit $code
 }
 
@@ -87,9 +92,16 @@ function check_add {
     fi
 
     echo -n $(pidof "$pid")" "
+
+    write_pidfile
 }
 
-trap 'terminate $?' 0 1 2 15
+function write_pidfile {
+    echo "$(allpids)" >$pidfile
+}
+
+trap 'terminate $?' EXIT
+trap 'terminate 1' HUP INT QUIT TERM
 
 # See if there is rabbitmq-server already running, if not, spawn one.
 if ! rabbitmqctl status >/dev/null 2>&1; then
@@ -114,29 +126,35 @@ export DJANGO_SETTINGS_MODULE=freeze_thaw_aws_test.settings
 
 env_setup
 
-echo -n "Initializing test database ... "
-(rm -f db.sqlite3 && \
-    $manage syncdb --noinput && \
-    $manage loaddata $test_dir/fixtures.yaml) >>$(logname freezr) 2>&1
-echo "done"
+if [ -n "$USE_EXISTING_FREEZR" ]; then
+    echo "Using existing freezr environment"
+else
+    echo -n "Initializing test database ... "
+    (rm -f db.sqlite3 && \
+	$manage syncdb --noinput && \
+	$manage loaddata $test_dir/fixtures.yaml) >>$(logname freezr) 2>&1
+    echo "done"
 
-# Celery ..
-echo -n "Starting celeryd ... "
-rm -f celeryd.pid
-$manage celeryd --concurrency 1 -B -l debug --pidfile celeryd.pid >>$(logname celeryd) 2>&1 &
-sleep 5; check_add "manage.py celeryd" "$(cat celeryd.pid)"
-echo "done"
+    # Celery ..
+    echo -n "Starting celeryd ... "
+    rm -f celeryd.pid
+    $manage celeryd --concurrency 1 -B -l debug \
+	--pidfile celeryd.pid >>$(logname celeryd) 2>&1 &
+    sleep 5; check_add "manage.py celeryd" "$(cat celeryd.pid)"
+    echo "done"
 
-export AWS_ACCESS_KEY_ID=$access_key
-export AWS_SECRET_ACCESS_KEY=$secret_key
-export AWS_DEFAULT_REGION=$region
+    export AWS_ACCESS_KEY_ID=$access_key
+    export AWS_SECRET_ACCESS_KEY=$secret_key
+    export AWS_DEFAULT_REGION=$region
 
-echo -n "Starting application server ... "
-$manage runserver 9000 >>$(logname freezr) 2>&1 &
-# can't use $! here, see https://code.djangoproject.com/ticket/19137
-# the extra sleep is *required*
-sleep 5; check_add "manage.py runserver 9000" 'runserver 9000'
-echo "done"
+    echo -n "Starting application server ... "
+    $manage runserver 9000 >>$(logname freezr) 2>&1 &
+
+    # can't use $! here, see https://code.djangoproject.com/ticket/19137
+    # the extra sleep is *required*
+    sleep 5; check_add "manage.py runserver 9000" 'runserver 9000'
+    echo "done"
+fi
 
 echo "Starting test suite, child pids are $(allpids)"
 
@@ -144,7 +162,18 @@ echo "Starting test suite, child pids are $(allpids)"
 # tests (if one fails, others might block etc.)
 if ! (env_setup && nosetests -x -v -w $test_dir); then
     echo "FAILURE: Tests failed"
-    exit 1
+    retval=1
+else
+    echo "SUCCESS: Tests passed"
+    retval=0
 fi
 
-exit 0
+if [[ $retval -ne 0 && -n "$KEEP_FAILED_FREEZR" ]]; then
+    echo "KEEP_FAILED_FREEZR specified, not killing pids $(allpids)"
+    pids=()
+elif [ -n "$USE_EXISTING_FREEZR" ]; then
+    echo "USE_EXISTING_FREEZR specified, not destroying environment"
+    pids=() #tautology, pids would be empty at this point anyway
+fi
+
+exit $retval
